Moved TemporaryDirectory and StreamCapture out of pints, into shared testing module. Closes #317.
